Configure log highlighting, add flag "-log-display-highlight"
With this change highlighting of logs displayed on stdout can be
configured. The default is to highlight using color, if stdout is a
terminal. This can be overridden with the -log-display-highlight flag -
to force color output, or force plain output.

An Enum is used, rather than a boolean to allow other highlighting
schemes (e.g. underline).

Fixes #169